% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sg_dummy}
\alias{sg_dummy}
\title{The function to store common parameters description}
\usage{
sg_dummy()
}
\arguments{
\item{abreaks}{Axis breaks function. Default is \code{scales::pretty_breaks(7)}}

\item{addOFV}{logical. If \code{TRUE}, information about OFV and AIC of the model will be added. Default is \code{TRUE}}

\item{addline}{logical. If \code{TRUE}, lines connecting observations of individual subjects will be added. Default is \code{TRUE}}

\item{alpha_i}{numeric. Transparency level (from 0 to 1) for points/lines. Default is 0.5}

\item{cap}{string. Plot caption.}

\item{cov_cols}{vector of characters. Name of the columns with covariates}

\item{col_i}{string. Column name for color}

\item{col_lab}{string. Label for color legend}

\item{dens}{Logical. If \code{TRUE}, plot histogram/density of residuals instead of scatter}

\item{ds_covs}{data.frame. The dataframe with covariates}

\item{eta_seq}{vector of strings. Character vector of parameter names to be plotted (e.g., \code{c("ka", "Cl")}). If \code{NULL}, all parameters be included. Default is \code{NULL}}

\item{f_scales}{one of \code{"fixed"}, \code{"free"}, \code{"free_x"}, \code{"free_y"}. User can specify whether the scales (x and y axes) should be fixed across all panels (\code{"fixed"}), free for each panel (\code{"free"}), or free only in one dimension (\code{"free_x"} or \code{"free_y"}). Default is \code{"fixed"}}

\item{facet_i}{string. Column name for facet}

\item{fpath_i}{string or sg-fit object. If the string is given, the path to \code{.Rdata} file with sg-fit object is expected}

\item{indiv}{logical. If \code{TRUE} uses individual predictions (\code{"IPRED"}); otherwise uses population predictions (\code{"PRED"}). Default is \code{TRUE}}

\item{lab_x}{string. X-axis label}

\item{lab_y}{string. Y-axis label}

\item{levels_discrete}{integer. Maximum unique values to consider a variable discrete. Default is 10}

\item{log_x}{logical. If \code{TRUE}, a logarithmic scale is applied to x-axis. Default is \code{FALSE}}

\item{log_y}{logical. If \code{TRUE}, a logarithmic scale is applied to y-axis. Default is \code{FALSE}}

\item{log_axes}{logical. If \code{TRUE}, a logarithmic scale is applied to all axes. Default is \code{FALSE}}

\item{max_x}{numeric. X ax maximum limit. Default is \code{NULL}}

\item{min_x}{numeric. X ax minimum limit. Default is \code{NULL}}

\item{max_y}{numeric. Y ax maximum limit. Default is \code{NULL}}

\item{min_y}{numeric. Y ax minimum limit. Default is \code{NULL}}

\item{n_bins}{integer. Number of bins to use in the histogram. Default is 30.}

\item{n_quantiles}{integer. Number of quantile groups for continuous variables in \code{col_i}. Default is 3}

\item{no_leg}{logical. If \code{TRUE}, no legend will be displayed. Default is \code{FALSE}}

\item{plot_type}{Character. Type of plot to produce:
\itemize{
\item \code{"DIST"} (default) - histogram of individual parameters,
\item \code{"QQ"} - QQ-plot of individual parameters
}}

\item{run_id}{integer. Tested model ID. Default is 1.}

\item{sc_factor}{numeric. Scaling factor for DV/PRED/IPRED values. Default is 1 (no scaling)}

\item{smooth}{logical. Add LOESS smooth line. Default is \code{TRUE}}

\item{tdist}{logical. If \code{TRUE}, overlay theoretical parameter distributions based on population mean and OMEGA matrix. Default is \code{TRUE}}

\item{tsld}{logical. If \code{TRUE}, uses time since last dose instead of time from first dose. Default is \code{FALSE}}
}
\description{
The function to store common parameters description
}
