% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-gof-par-dist.R
\name{sg_gof_par_dist}
\alias{sg_gof_par_dist}
\title{Plot Distributions of Individual Parameters with Optional Theoretical Density}
\usage{
sg_gof_par_dist(
  fpath_i,
  eta_seq = NULL,
  n_bins = 30,
  tdist = T,
  plot_type = "DIST"
)
}
\arguments{
\item{fpath_i}{string or sg-fit object. If the string is given, the path to \code{.Rdata} file with sg-fit object is expected}

\item{eta_seq}{vector of strings. Character vector of parameter names to be plotted (e.g., \code{c("ka", "Cl")}). If \code{NULL}, all parameters be included. Default is \code{NULL}}

\item{n_bins}{integer. Number of bins to use in the histogram. Default is 30.}

\item{tdist}{logical. If \code{TRUE}, overlay theoretical parameter distributions based on population mean and OMEGA matrix. Default is \code{TRUE}}

\item{plot_type}{Character. Type of plot to produce:
\itemize{
\item \code{"DIST"} (default) - histogram of individual parameters,
\item \code{"QQ"} - QQ-plot of individual parameters
}}
}
\value{
A \code{ggplot} object containing histogram(s) of individual parameter distributions, optionally overlaid with theoretical densities.
}
\description{
This function creates histograms of individual parameter estimates,
with optional overlay of the theoretical distribution based on the population parameters and OMEGA matrix.
}
\details{
The function visualizes the distribution of post hoc individual parameter estimates (from \verb{$PATAB}) and compares them with the expected population-level variability assuming log-normal distribution.
Shrinkage values are calculated based on \code{ETAshrinkage_sd} from \verb{$SUMTAB} and included in facet labels.

Theoretical distributions are generated by sampling from a multivariate normal distribution (mean zero, covariance from \verb{$OMEGAMAT}) and transforming to the log-normal scale using the typical values (\code{VALUE} column of \verb{$SUMTAB}).
}
\examples{
\dontrun{
# Basic usage
sg_gof_par_dist(fpath_i = "PK.RData")

# Only for specific parameters
sg_gof_par_dist(fpath_i = "~/Simurg_function/Warfarin_PK.RData", eta_seq = c("ka", "Cl"))
}
}
