% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-gof-res.R
\name{sg_gof_res}
\alias{sg_gof_res}
\title{Create residual diagnostic plots (WRES, IWRES, CWRES, etc.)}
\usage{
sg_gof_res(
  fpath_i,
  res_type,
  cov_cols = NULL,
  indiv = TRUE,
  vs_time = TRUE,
  addline = TRUE,
  alpha_i = 0.5,
  smooth = TRUE,
  log_x = FALSE,
  sc_factor = 1,
  abreaks = scales::pretty_breaks(7),
  lab_y = NULL,
  lab_x = NULL,
  col_i = NULL,
  col_lab = NULL,
  facet_i = NULL,
  f_scales = "fixed",
  dens = FALSE,
  n_bins = 50,
  ymin = NA,
  ymax = NA,
  xmin = NA,
  xmax = NA,
  no_leg = FALSE,
  n_quantiles = 3,
  levels_discrete = 10
)
}
\arguments{
\item{fpath_i}{string or sg-fit object. If the string is given, the path to \code{.Rdata} file with sg-fit object is expected}

\item{res_type}{Character. Name of residual column to plot (e.g., \code{"WRES"}, \code{"IWRES"}, \code{"CWRES"})}

\item{cov_cols}{vector of characters. Name of the columns with covariates}

\item{indiv}{logical. If \code{TRUE} uses individual predictions (\code{"IPRED"}); otherwise uses population predictions (\code{"PRED"}). Default is \code{TRUE}}

\item{vs_time}{Logical. If \code{TRUE}, plot residuals vs TIME, else vs predictions (IPRED/PRED)}

\item{addline}{Logical. Add connecting lines per subject. Default \code{TRUE}}

\item{alpha_i}{numeric. Transparency level (from 0 to 1) for points/lines. Default is 0.5}

\item{smooth}{logical. Add LOESS smooth line. Default is \code{TRUE}}

\item{log_x}{logical. If \code{TRUE}, a logarithmic scale is applied to x-axis. Default is \code{FALSE}}

\item{sc_factor}{numeric. Scaling factor for DV/PRED/IPRED values. Default is 1 (no scaling)}

\item{abreaks}{Axis breaks function. Default is \code{scales::pretty_breaks(7)}}

\item{lab_y}{string. Y-axis label}

\item{lab_x}{string. X-axis label}

\item{col_i}{string. Column name for color}

\item{col_lab}{string. Label for color legend}

\item{facet_i}{string. Column name for facet}

\item{f_scales}{one of \code{"fixed"}, \code{"free"}, \code{"free_x"}, \code{"free_y"}. User can specify whether the scales (x and y axes) should be fixed across all panels (\code{"fixed"}), free for each panel (\code{"free"}), or free only in one dimension (\code{"free_x"} or \code{"free_y"}). Default is \code{"fixed"}}

\item{dens}{Logical. If \code{TRUE}, plot histogram/density of residuals instead of scatter}

\item{n_bins}{integer. Number of bins to use in the histogram. Default is 30.}

\item{no_leg}{logical. If \code{TRUE}, no legend will be displayed. Default is \code{FALSE}}

\item{n_quantiles}{integer. Number of quantile groups for continuous variables in \code{col_i}. Default is 3}

\item{levels_discrete}{Integer. Threshold for treating covariates as discrete. Default 10}
}
\value{
A \code{ggplot2} object
}
\description{
Generates residual diagnostic plots either versus time/predictions
or as residual distributions. Supports faceting, covariate coloring,
quantile binning for continuous covariates, and optional smoothing.
}
\examples{
\donttest{
# Basic residuals vs time
sg_gof_res(
  fpath_i = paste0("functions/nlme/3.4-sg-gof-res/simurg-object/Warfarin_PK.RData"),
  res_type = "IWRES",
  vs_time = TRUE
)
# IWRES dist
sg_gof_res(
  fpath_i = paste0("functions/nlme/3.4-sg-gof-res/simurg-object/Warfarin_PK.RData"),
  res_type = "IWRES",
  cov_cols = "SEX",
  dens = T,
  lab_x = "IWRES",
  facet_i = "SEX",
  addline = T) + labs(subtitle = "Individual")
}

}
