% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-gof-res-dist.R
\name{sg_gof_res_dist}
\alias{sg_gof_res_dist}
\title{Plot distribution or QQ-plot of residuals from NONMEM/Simurg run}
\usage{
sg_gof_res_dist(
  fpath_i,
  res_type = "RES",
  n_bins = 30,
  ndist = T,
  plot_type = "DIST"
)
}
\arguments{
\item{fpath_i}{string or sg-fit object. If the string is given, the path to \code{.Rdata} file with sg-fit object is expected}

\item{res_type}{Character vector. One or several types of residuals to plot.
Values must correspond to column name(s) in \code{smrg_obj$SDTAB}, e.g.
\code{"RES"}, \code{"IWRES"}, \code{"IRES"}. If multiple residual types are provided,
plots will be generated for each of them.}

\item{n_bins}{integer. Number of bins to use in the histogram. Default is 30.}

\item{ndist}{Logical. If \code{TRUE}, overlays the corresponding normal density
curve on the histogram (default: TRUE).}

\item{plot_type}{Character. Type of plot to produce:
\itemize{
\item \code{"DIST"} (default) - histogram of individual parameters,
\item \code{"QQ"} - QQ-plot of individual parameters
}}
}
\value{
A \code{ggplot} object.
}
\description{
This function visualizes residuals from estimation results
stored in a Simurg output object. The residuals can be plotted either as
histograms with overlaid normal density curves, or as QQ-plots to assess
normality.
}
\examples{
\dontrun{
# Plot the distribution of individual weighted residuals (IWRES)
# as histograms with overlaid normal curves for each DVID
sg_gof_res_dist(fpath_i = "PK.RData", res_type = "IWRES")

# Generate QQ-plots to visually assess normality of standard residuals (RES)
# stratified by DVID
sg_gof_res_dist(fpath_i = "PK.RData", res_type = "RES", plot_type = "QQ")

# Multiple residual types can be specified (e.g., IWRES and IRES);
# the function will produce plots for each residual type across all DVID groups
sg_gof_res_dist(fpath_i = "PK.RData", res_type = c("IWRES", "IRES"))
}
}
