% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-gof-tp.R
\name{sg_gof_tp}
\alias{sg_gof_tp}
\title{Plot time profiles of the fitted data}
\usage{
sg_gof_tp(
  fpath_i,
  filt = "T",
  tsld = F,
  f_scales = "free",
  log_y = F,
  lab_x = "Time since first dose, h",
  lab_y = "Plasma concentration, mmol/L",
  cap =
    "empty circles - observed data\\nsolid lines with point - individual predictions\\ndashed grey lines with point - population predictions"
)
}
\arguments{
\item{fpath_i}{string or sg-fit object. If the string is given, the path to \code{.Rdata} file with sg-fit object is expected}

\item{tsld}{logical. If \code{TRUE}, uses time since last dose instead of time from first dose. Default is \code{FALSE}}

\item{f_scales}{one of \code{"fixed"}, \code{"free"}, \code{"free_x"}, \code{"free_y"}. User can specify whether the scales (x and y axes) should be fixed across all panels (\code{"fixed"}), free for each panel (\code{"free"}), or free only in one dimension (\code{"free_x"} or \code{"free_y"}). Default is \code{"fixed"}}

\item{log_y}{logical. If \code{TRUE}, a logarithmic scale is applied to y-axis. Default is \code{FALSE}}

\item{lab_x}{string. X-ax label. Default is "Time since first dose, h"}

\item{lab_y}{string. Y-ax label. Default is "Plasma concentration, mmol/L"}

\item{cap}{string. Plot caption. Default is "empty circles - observed data\n solid lines with point - individual predictions\n dashed grey lines with point - population predictions"}
}
\value{
A list of plots with predicted time profiles, faceted by id
}
\description{
Plot time profiles of the fitted data
}
\examples{
\dontrun{
fpath_i <- system.file("extdata", "simurg_object", "Warfarin_PK.RData", package = "SimuRg"))
plot_list <- sg_gof_tp(fpath_i)
print(plot_list[[1]])
}
}
