% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-modcomp.R
\name{sg_modcomp}
\alias{sg_modcomp}
\title{Returns dataframe with summary statistics for model comparison}
\usage{
sg_modcomp(fpath_i, run_id = 1)
}
\arguments{
\item{fpath_i}{string or sg-fit object. If the string is given, the path to \code{.Rdata} file with sg-fit object is expected}

\item{run_id}{integer. Tested model ID. Default is 1.}
}
\value{
A data.frame with model summary metrics for futher comparison
}
\description{
Returns dataframe with summary statistics for model comparison
}
\examples{
\dontrun{
fpath_i <- system.file("extdata", "simurg_object", "Warfarin_PK.RData", package = "SimuRg")
sum_tab <- sg_modcomp(fpath_i, 3)
print(sum_tab)
}
}
