% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg-parsum.R
\name{sg_parsum}
\alias{sg_parsum}
\title{Extract parameter summary (theta, eta, SE, RSE, CI, CV, shrinkage)}
\usage{
sg_parsum(fpath_i, addOFV = T)
}
\arguments{
\item{fpath_i}{string or sg-fit object. If the string is given, the path to \code{.Rdata} file with sg-fit object is expected}

\item{addOFV}{logical. If \code{TRUE}, information about OFV and AIC of the model will be added. Default is \code{TRUE}}

\item{cap}{string. Plot caption. Default is "empty circles - observed data\n solid lines with point - individual predictions\n dashed grey lines with point - population predictions"}
}
\value{
A table with parameter summary
}
\description{
Extract parameter summary (theta, eta, SE, RSE, CI, CV, shrinkage)
}
\examples{
\dontrun{
fpath_i <- system.file("extdata", "simurg_object", "Warfarin_PK.RData", package = "SimuRg"))
sum_tab <- sg_parsum(fpath_i)
print(sum_tab)
}
}
